requirements.txt
silero-vad==5.1
unsilence==1.0.9
--extra-index-url https://download.pytorch.org/whl/cu118
altgraph==0.17.4
asgiref==3.8.1
asttokens==2.4.1
bottle==0.13.0
certifi==2024.8.30
cffi==1.17.1
clr-loader==0.2.6
colorama==0.4.6
coloredlogs==15.0.1
commonmark==0.9.1
decorator==5.1.1
distlib==0.3.8
Django==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.15.4
flatbuffers==24.3.25
fsspec==2024.9.0
humanfriendly==10.0
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.3
numpy==2.1.1
onnxruntime==1.19.2
packaging==24.1
parso==0.8.4
pefile==2024.8.26
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==5.28.1
proxy_tools==0.1.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyreadline3==3.4.3
python-dotenv==1.0.1
pythonnet==3.0.3
pywebview==5.2
pywin32-ctypes==0.2.3
rich==10.10.0
six==1.16.0
soundfile==0.12.1
sqlparse==0.5.1
stack-data==0.6.3
sv-ttk==2.6.0
sympy==1.13.2
torch==2.4.1+cu118
torchaudio==2.4.1+cu118
torchvision==0.19.1+cu118
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
virtualenv==20.26.3
wcwidth==0.2.13
